requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Segmentation Models
segmentation-models-pytorch>=0.3.3

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Data Handling
pandas>=2.0.0
h5py>=3.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Progress Bars
tqdm>=4.65.0

# Configuration
PyYAML>=6.0
hydra-core>=1.3.0

# Metrics and Evaluation
torchmetrics>=1.0.0

# Model Export
onnx>=1.14.0
onnxruntime>=1.15.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.13.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Web Application
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-aggrid>=0.3.4

# Development Tools
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0

# CUDA support (optional, install manually based on system)
# torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118